mudando o get do card para o menu
