+++ pokedex/src/Components/Buttons/GoToDetails.js
@@ -1,0 +1,19 @@
+import { goToDetails } from '../../Router/Coordinator'
+import { useHistory } from 'react-router-dom'
+import styled from 'styled-components'
+
+const Button = styled.button`
+`
+
+const GoToDetails = () => {
+
+    const history = useHistory()
+
+    
+
+    return (
+        <Button onClick={() => goToDetails(history)}>Ver Detalhes</Button>
+    )
+}
+
+export default GoToDetails
